Move getting the release note data to shared code
